data-quality-governance-pipeline/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
plotly>=5.0.0
scikit-learn>=1.0.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
requests>=2.25.0
psutil>=5.8.0
pytest>=6.0.0
pytest-cov>=2.12.0
